requirements.txt
boto==2.39.0
Django==1.9.1
django-grappelli==2.7.3
django-storages==1.1.8
Pillow==3.1.0
psycopg2==2.6.1
wheel==0.24.0
